actually pull Base* out of Swapchain
