+++ sdk/engine/source/assets/CubeMapAssetManager.cpp
@@ -2,7 +2,7 @@
 #include "assets/AssetUtilities.h"
 #include "graphics/Base.h"
 #include "graphics/Graphics.h"
-#include "graphics/Initializers.h"
+#include "graphics/vk/Initializers.h"
 #include "graphics/resources/ShaderResourceService.h"
 
 #include <cassert>
