+++ sdk/engine/source/assets/TextureAssetManager.cpp
@@ -1,7 +1,7 @@
 #include "TextureAssetManager.h"
 #include "graphics/Base.h"
 #include "graphics/Graphics.h"
-#include "graphics/Initializers.h"
+#include "graphics/vk/Initializers.h"
 #include "graphics/resources/TextureManager.h"
 
 #define STB_IMAGE_IMPLEMENTATION
