+++ sdk/engine/source/graphics/Base.cpp
@@ -1,9 +1,10 @@
 #include "Base.h"
-#include "graphics/Initializers.h"
+#include "graphics/vk/Initializers.h"
 #include "graphics/Commands.h"
 #include "stb/stb_image.h"
 #include "debug/NcError.h"
 #include "config/Config.h"
+#include "vk/Swapchain.h"
 
 #include <set>
 #include <string>
@@ -522,120 +523,5 @@ namespace nc::graphics
 			}
 		}
 		throw NcError("Could not find a matching depth format");
-    }
-
-    const SwapChainSupportDetails Base::QuerySwapChainSupport(const vk::PhysicalDevice& device, const vk::SurfaceKHR& surface) const
-    {
-        SwapChainSupportDetails details;
-        if (device.getSurfaceCapabilitiesKHR(surface, &details.capabilities) != vk::Result::eSuccess)
-        {
-            throw NcError("SwapChain::QuerySwapChainSupport() - Could not enumerate surface capabilities.");
-        }
-
-        uint32_t formatCount;
-        if (device.getSurfaceFormatsKHR(surface, &formatCount, nullptr) != vk::Result::eSuccess)
-        {
-            throw NcError("SwapChain::QuerySwapChainSupport() - Could not enumerate surface formats.");
-        }
-
-        if (formatCount != 0)
-        {
-            details.formats.resize(formatCount);
-            if (device.getSurfaceFormatsKHR(surface, &formatCount, details.formats.data()) != vk::Result::eSuccess)
-            {
-                throw NcError("SwapChain::QuerySwapChainSupport() - Could not enumerate surface formats.");
-            }
-        }
-
-        uint32_t presentModeCount;
-        if (device.getSurfacePresentModesKHR(surface, &presentModeCount, nullptr) != vk::Result::eSuccess)
-        {
-            throw NcError("SwapChain::QuerySwapChainSupport() - Could not enumerate surface present modes.");
-        }
-
-        if (presentModeCount != 0)
-        {
-            details.presentModes.resize(presentModeCount);
-            if (device.getSurfacePresentModesKHR(surface, &presentModeCount, details.presentModes.data()) != vk::Result::eSuccess)
-            {
-                throw NcError("SwapChain::QuerySwapChainSupport() - Could not enumerate surface present modes.");
-            }
-        } 
-        return details;
-    }
-
-    QueueFamilyIndices::QueueFamilyIndices(const vk::PhysicalDevice& device, const vk::SurfaceKHR& surface)
-    {
-        uint32_t queueFamilyCount = 0;
-        device.getQueueFamilyProperties(&queueFamilyCount, nullptr);
-
-        std::vector<vk::QueueFamilyProperties> queueFamilies(queueFamilyCount);
-        device.getQueueFamilyProperties(&queueFamilyCount, queueFamilies.data());
-        
-        uint32_t i = 0;
-        for (const auto& queueFamily : queueFamilies)
-        {
-            vk::Bool32 presentSupport = false;
-            if (device.getSurfaceSupportKHR(i, surface, &presentSupport) != vk::Result::eSuccess)
-            {
-                throw NcError("Could not get surface support KHR");
-            }
-
-            if (queueFamily.queueFlags & vk::QueueFlagBits::eGraphics)
-            {
-                m_graphicsFamily = i;
-            }
-
-            if (presentSupport)
-            {
-                m_presentFamily = i;
-            }
-
-            if (IsComplete())
-            {
-                break;
-            }
-
-            ++i;
-        }
-
-        if (!m_graphicsFamily.has_value())
-        {
-            throw NcError("No graphics queue family found on device.");
-        }
-
-        if (!m_presentFamily.has_value())
-        {
-            throw NcError("No graphics queue family found on device.");
-        }
-
-        m_isSeparatePresentQueue = m_presentFamily != m_graphicsFamily ? true : false;
-    }
-
-    bool QueueFamilyIndices::IsComplete() const
-    {
-        return m_graphicsFamily.has_value() && m_presentFamily.has_value();
-    }
-
-    bool QueueFamilyIndices::IsSeparatePresentQueue() const
-    {
-        return m_isSeparatePresentQueue;
-    }
-
-    uint32_t QueueFamilyIndices::GetQueueFamilyIndex(QueueFamilyType type) const
-    {
-        if (!IsComplete())
-        {
-            throw NcError("QueueFamilyIndices::GetQueueFamilyIndex() - QueueFamilies incomplete.");
-        }
-
-        switch (type)
-        {
-            case QueueFamilyType::GraphicsFamily:
-                return m_graphicsFamily.value();
-            case QueueFamilyType::PresentFamily:
-                return m_presentFamily.value();
-        }
-        throw NcError("QueueFamilyIndices::GetQueueFamilyIndex() - Chosen queue not present.");
     }
 }
