+++ sdk/engine/source/graphics/Base.h
@@ -7,6 +7,7 @@
 #endif
 #include "imgui/imgui_impl_vulkan.h"
 #include "GpuAllocator.h"
+#include "vk/VkTypes.h"
 
 #include <optional>
 #include <unordered_map>
@@ -20,37 +21,6 @@ namespace nc::graphics
     class Commands;
     class TexturesData;
     struct Vertex;
-
-    // How many frames can be rendered concurrently.
-    // Each frame requires its own pair of semaphores.
-    constexpr uint32_t MaxFramesInFlight = 2u;
-
-    enum class QueueFamilyType : uint8_t
-    {
-        GraphicsFamily,
-        PresentFamily
-    };
-
-    struct SwapChainSupportDetails
-    {
-        vk::SurfaceCapabilitiesKHR capabilities;
-        std::vector<vk::SurfaceFormatKHR> formats;
-        std::vector<vk::PresentModeKHR> presentModes;
-    };
-
-    class QueueFamilyIndices
-    {
-        public:
-            QueueFamilyIndices(const vk::PhysicalDevice& device, const vk::SurfaceKHR& surface);
-            bool IsComplete() const;
-            bool IsSeparatePresentQueue() const;
-            uint32_t GetQueueFamilyIndex(QueueFamilyType type) const;
-
-        private:
-            std::optional<uint32_t> m_graphicsFamily;
-            std::optional<uint32_t> m_presentFamily;
-            bool m_isSeparatePresentQueue;
-    };
 
     class Base
     {
@@ -70,7 +40,7 @@ namespace nc::graphics
             vk::UniqueSampler CreateTextureSampler();
 
             void FreeCommandBuffers(std::vector<vk::CommandBuffer>* commandBuffers) noexcept;
-            const SwapChainSupportDetails QuerySwapChainSupport(const vk::PhysicalDevice& device, const vk::SurfaceKHR& surface) const;
+            //const SwapChainSupportDetails QuerySwapChainSupport(const vk::PhysicalDevice& device, const vk::SurfaceKHR& surface) const;
             void QueryDepthFormatSupport();
             void InitializeImgui(const vk::RenderPass& defaultPass);
             void TransitionImageLayout(vk::Image image, vk::ImageLayout oldLayout, uint32_t layerCount, vk::ImageLayout newLayout);
