+++ sdk/engine/source/graphics/Renderer.cpp
@@ -7,12 +7,12 @@
 #include "graphics/Base.h"
 #include "graphics/Commands.h"
 #include "graphics/Graphics.h"
-#include "graphics/Swapchain.h"
 #include "graphics/techniques/EnvironmentTechnique.h"
 #include "graphics/techniques/ParticleTechnique.h"
 #include "graphics/techniques/PbrTechnique.h"
 #include "graphics/techniques/UiTechnique.h"
 #include "graphics/techniques/ShadowMappingTechnique.h"
+#include "graphics/vk/Swapchain.h"
 #include "optick/optick.h"
 #include "PerFrameRenderState.h"
 #include "resources/RenderTarget.h"
