+++ sdk/engine/source/graphics/resources/CubeMapManager.cpp
@@ -1,7 +1,7 @@
 #include "CubeMapManager.h"
 #include "graphics/Base.h"
 #include "graphics/Graphics.h"
-#include "graphics/Initializers.h"
+#include "graphics/vk/Initializers.h"
 #include "debug/Utils.h"
 
 #include <array>
