+++ sdk/engine/source/graphics/resources/ObjectDataManager.cpp
@@ -1,5 +1,5 @@
 #include "ObjectDataManager.h"
-#include "graphics/Initializers.h"
+#include "graphics/vk/Initializers.h"
 
 namespace nc::graphics
 {
