+++ sdk/engine/source/graphics/resources/PointLightManager.cpp
@@ -1,5 +1,5 @@
 #include "PointLightManager.h"
-#include "graphics/Initializers.h"
+#include "graphics/vk/Initializers.h"
 
 namespace nc::graphics
 {
