+++ sdk/engine/source/graphics/resources/RenderPassManager.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "graphics/Initializers.h"
+#include "graphics/vk/Initializers.h"
 #include "graphics/techniques/ITechnique.h"
 
 #include "vulkan/vk_mem_alloc.hpp"
