+++ sdk/engine/source/graphics/resources/ShaderDescriptorSets.cpp
@@ -1,6 +1,6 @@
 #include "ShaderDescriptorSets.h"
 #include "graphics/Base.h"
-#include "graphics/Initializers.h"
+#include "graphics/vk/Initializers.h"
 
 namespace
 {
