+++ sdk/engine/source/graphics/resources/ShadowMapManager.cpp
@@ -1,7 +1,7 @@
 #include "ShadowMapManager.h"
 #include "graphics/Base.h"
 #include "graphics/Graphics.h"
-#include "graphics/Initializers.h"
+#include "graphics/vk/Initializers.h"
 
 #include <vector>
