+++ sdk/engine/source/graphics/resources/TextureManager.cpp
@@ -1,5 +1,5 @@
 #include "TextureManager.h"
-#include "graphics/Initializers.h"
+#include "graphics/vk/Initializers.h"
 #include "debug/Utils.h"
 
 #include <array>
