+++ sdk/engine/source/graphics/techniques/EnvironmentTechnique.cpp
@@ -4,15 +4,15 @@
 #include "config/Config.h"
 #include "graphics/Graphics.h"
 #include "graphics/Commands.h"
-#include "graphics/Initializers.h"
+#include "graphics/vk/Initializers.h"
 #include "graphics/ShaderUtilities.h"
 #include "graphics/PerFrameRenderState.h"
-#include "graphics/Swapchain.h"
 #include "graphics/Base.h"
 #include "graphics/VertexDescriptions.h"
 #include "graphics/resources/Environment.h"
 #include "graphics/resources/ImmutableBuffer.h"
 #include "graphics/resources/ShaderResourceServices.h"
+#include "graphics/vk/Swapchain.h"
 #include "optick/optick.h"
 
 namespace nc::graphics
