+++ sdk/engine/source/graphics/techniques/ParticleTechnique.cpp
@@ -5,7 +5,7 @@
 #include "graphics/Base.h"
 #include "graphics/Commands.h"
 #include "graphics/Graphics.h"
-#include "graphics/Initializers.h"
+#include "graphics/vk/Initializers.h"
 #include "graphics/PerFrameRenderState.h"
 #include "graphics/ShaderUtilities.h"
 #include "graphics/resources/ImmutableBuffer.h"
