+++ sdk/engine/source/graphics/techniques/PbrTechnique.cpp
@@ -3,7 +3,7 @@
 #include "config/Config.h"
 #include "graphics/Graphics.h"
 #include "graphics/Commands.h"
-#include "graphics/Initializers.h"
+#include "graphics/vk/Initializers.h"
 #include "graphics/ShaderUtilities.h"
 #include "graphics/PerFrameRenderState.h"
 #include "graphics/Base.h"
