+++ sdk/engine/source/graphics/techniques/ShadowMappingTechnique.cpp
@@ -4,7 +4,7 @@
 #include "ecs/Registry.h"
 #include "graphics/Base.h"
 #include "graphics/Graphics.h"
-#include "graphics/Initializers.h"
+#include "graphics/vk/Initializers.h"
 #include "graphics/resources/ShaderResourceServices.h"
 #include "graphics/ShaderUtilities.h"
 #include "graphics/VertexDescriptions.h"
