+++ sdk/engine/source/graphics/techniques/UiTechnique.cpp
@@ -4,11 +4,11 @@
 #include "ecs/Registry.h"
 #include "graphics/Base.h"
 #include "graphics/Graphics.h"
-#include "graphics/Initializers.h"
+#include "graphics/vk/Initializers.h"
 #include "graphics/resources/ImmutableBuffer.h"
 #include "graphics/ShaderUtilities.h"
-#include "graphics/Swapchain.h"
 #include "graphics/VertexDescriptions.h"
+#include "graphics/vk/Swapchain.h"
 
 namespace nc::graphics
 {
