+++ sdk/engine/source/graphics/techniques/WireframeTechnique.cpp
@@ -7,11 +7,11 @@
 #include "graphics/Base.h"
 #include "graphics/DebugRenderer.h"
 #include "graphics/Graphics.h"
-#include "graphics/Initializers.h"
+#include "graphics/vk/Initializers.h"
 #include "graphics/resources/ImmutableBuffer.h"
 #include "graphics/ShaderUtilities.h"
-#include "graphics/Swapchain.h"
 #include "graphics/VertexDescriptions.h"
+#include "graphics/vk/Swapchain.h"
 
 namespace nc::graphics
 {
