+++ sdk/engine/source/graphics/vk/CMakeLists.txt
@@ -1,4 +1,6 @@
 target_sources(${NC_ENGINE_LIB}
     PRIVATE
+        Initializers.cpp
+        Swapchain.cpp
         VkTypes.cpp
 )
