+++ sdk/engine/source/graphics/vk/VkTypes.h
@@ -6,11 +6,32 @@
 #include "vulkan/vk_mem_alloc.hpp"
 #include <array>
 #include <memory>
+#include <optional>
 #include <span>
 #include <vector>
 
 namespace nc::graphics
 {
+enum class QueueFamilyType : uint8_t
+{
+    GraphicsFamily,
+    PresentFamily
+};
+
+class QueueFamilyIndices
+{
+    public:
+        QueueFamilyIndices(const vk::PhysicalDevice& device, const vk::SurfaceKHR& surface);
+        bool IsComplete() const;
+        bool IsSeparatePresentQueue() const;
+        uint32_t GetQueueFamilyIndex(QueueFamilyType type) const;
+
+    private:
+        std::optional<uint32_t> m_graphicsFamily;
+        std::optional<uint32_t> m_presentFamily;
+        bool m_isSeparatePresentQueue;
+};
+
 enum class ShaderStage : uint8_t
 {
     Vertex,
